requirements.txt
altair
numpy
pandas
pydeck
streamlit
openai
python-dotenv
requests
pdfkit
beautifulsoup4